% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dycor.R
\name{ax.ay}
\alias{ax.ay}
\title{Compute correlation between a person's X value and their Y value,
while controlling for dyadic dependence in the data.}
\usage{
ax.ay(.data, ax, ay, bx, by)
}
\arguments{
\item{.data}{Data frame or the like containing the columns below}

\item{ax}{Column name that contains a person A's score on variable X}

\item{ay}{Column name that contains a person A's score on variable Y}

\item{bx}{Column name that contains a person B's score on variable X}

\item{by}{Column name that contains a person B's score on variable Y}
}
\value{
Correlation, p value, and adjusted sample size (N star)
}
\description{
We start from a data table where we have scores of person A and B on variables X and Y.
Data format requirement: all dyad members need to be present in the data
once as A and once as B. That is, each dyad corresponds to two, not one, row.
The question asked here is whether a person's X value predicts their own Y value.
This is within-dyad, but it takes into account potential dependence between
values of the same dyad.
The question asked here is whether a person's X value predicts their own Y value (ax => ay).
This is within-dyad of course, but it takes into account potential dependence between
values of the same dyad.
Scientific Reference:
Griffin, D., & Gonzalez, R. (1995). Correlational analysis of dyad-level data in the exchangeable case. Psychological
Bulletin, 118(3), 430–439. https://doi.org/10.1037/0033-2909.118.3.430
}
