% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dycor.R
\name{cor.structure}
\alias{cor.structure}
\title{Helper function that computes some basic correlations}
\usage{
cor.structure(ax, ay, bx, by)
}
\arguments{
\item{ax}{Vector containing a person A's score on variable X}

\item{ay}{Vector containing a person A's score on variable Y}

\item{bx}{Vector containing a person B's score on variable X}

\item{by}{Vector containing a person B's score on variable Y}
}
\value{
A list of correlations (e.g. raxbx = correlation between a person's X value and their partner's Y value)
}
\description{
Computes a set of basic correlations between
variables, both within a person and between people in the dyads.
These are then used to compute more complex correlations and significance levels.
}
