% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dycor.R
\name{z.to.p}
\alias{z.to.p}
\title{Convert z value into two-tailed p value}
\usage{
z.to.p(z)
}
\arguments{
\item{z}{Z value to be converted into p}
}
\value{
Corresponding p value
}
\description{
Convert z value into two-tailed p value
}
